test: Move toMatchTaskWithPath() & toBeValid() to separate file
